requirements.txt
python-telegram-bot==21.4
python-dotenv==1.0.1
pandas==2.2.2
openpyxl==3.1.5
